+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 ## Unreleased
+- Add `Instance` value object
 
 ## 7.7.0 - 2026-05-18
 - Add `Environment` and `Tier` value objects